Finished adding punch, bug fix and more.
The punch attack is now properly added into the game, dealing 20 damage to targets and 10 to the attacker (which is intended), if it hits something that can take damage, then it deals the damage and heals the attacker for 60% of the damage dealt to the target (here it's 12hp since the attack deals 20 damage). Also simplified the look of the hands to make it easier to animate, added 3 entries to the input map for 3 future attacks (those attacks aren't in the game yet though, the entries of the input map are simply here for when I will eventually add the attacks in the game), added a health monitor which displays the maximum amount of health the player can have and how much health does the player currently have, added a dummy for testing out attacks, decreased the cooldown on the punch from 2 seconds to 1 second, added a "take_damage" function which makes the player take a certain amount of damage and will check whether the player is invincible or not (if yes then nothing happens), whether the source of the damage is internal (caused by the player itself attacking) or external (not caused by the player), and then it will deal the damage (for now the source doesn't change anything but later on I plan on making it play different animations depending on the source of damage), added a "heal" function which simply adds health to the player, and if the health after the heal surpasses the max health, then the health is set back to the max health, and added a "deal_damage" function which checks if the target has a health property and if it can take damage, then checks if the target is invincible (if it is then it will call the target's "take_damage" method but that method checks for invincibility anyway), and if it's not, then it will call the target's "take_damage" method AND the player will heal for 60% of the damage dealt to the target (so if 20 damage is dealt, then the player will take 10 damage but heal 12hp). Also if the "camera_up" or "camera_down" input gets pressed quickly enough, the camera will stay up (or down), which is normal, however the abnormal part is that this would stack up, causing the camera to rotate way too far vertically, that part got fixed.